samples/06-ellar-and-django-orm/requirements.txt
Django
ellar-cli
ellar-django-module
